chore: relocate torch_multi_arange
