+++ .pre-commit-config.yaml
@@ -1031,7 +1031,6 @@ common-files: &common_files |
         tests/unittest/_torch/ray_orchestrator/single_gpu/test_cache_transceiver_comm.py |
         tests/unittest/_torch/sampler/test_beam_search.py |
         tests/unittest/_torch/sampler/test_best_of_n.py |
-        tests/unittest/_torch/sampler/test_torch_multi_arange.py |
         tests/unittest/_torch/sampler/test_trtllm_sampler.py |
         tests/unittest/_torch/speculative/test_draft_target.py |
         tests/unittest/_torch/speculative/test_draft_token_tree_sampling.py |
@@ -1046,6 +1045,7 @@ common-files: &common_files |
         tests/unittest/_torch/speculative/test_torch_rejection_sampling.py |
         tests/unittest/_torch/speculative/test_user_provided.py |
         tests/unittest/_torch/test_connector.py |
+        tests/unittest/_torch/test_torch_multi_arange.py |
         tests/unittest/_torch/thop/parallel/deep_gemm_tests.py |
         tests/unittest/_torch/thop/parallel/test_causal_conv1d_op.py |
         tests/unittest/_torch/thop/parallel/test_cublas_mm.py |
@@ -2406,7 +2406,6 @@ legacy-files: &legacy_files |
         tests/unittest/_torch/ray_orchestrator/single_gpu/test_cache_transceiver_comm.py |
         tests/unittest/_torch/sampler/test_beam_search.py |
         tests/unittest/_torch/sampler/test_best_of_n.py |
-        tests/unittest/_torch/sampler/test_torch_multi_arange.py |
         tests/unittest/_torch/sampler/test_trtllm_sampler.py |
         tests/unittest/_torch/speculative/test_draft_target.py |
         tests/unittest/_torch/speculative/test_draft_token_tree_sampling.py |
@@ -2421,6 +2420,7 @@ legacy-files: &legacy_files |
         tests/unittest/_torch/speculative/test_torch_rejection_sampling.py |
         tests/unittest/_torch/speculative/test_user_provided.py |
         tests/unittest/_torch/test_connector.py |
+        tests/unittest/_torch/test_torch_multi_arange.py |
         tests/unittest/_torch/thop/parallel/deep_gemm_tests.py |
         tests/unittest/_torch/thop/parallel/test_causal_conv1d_op.py |
         tests/unittest/_torch/thop/parallel/test_cublas_mm.py |
