+++ tensorrt_llm/_torch/pyexecutor/sampler.py
@@ -83,6 +83,7 @@ from tensorrt_llm.sampling_params import (
 from ..flashinfer_utils import IS_FLASHINFER_AVAILABLE
 from ..speculative.interface import get_force_num_accepted_tokens
 from ..speculative.spec_tree_manager import SpecTreeManager
+from ..utils import torch_multi_arange
 from .finish_reason import FinishedState
 from .llm_request import LlmRequest, LlmRequestState, get_draft_token_length
 from .resource_manager import ResourceManager, ResourceManagerType
@@ -100,7 +101,6 @@ from .sampling_utils import (
     resolve_sampling_strategy,
     sample,
     sample_rejected,
-    torch_multi_arange,
 )
 from .scheduler import ScheduledRequests
